Develop

+++ src/schemas/movies.schemas.ts
@@ -3,15 +3,15 @@ import { z } from "zod";
 const movieSchema = z.object({
     name: z.string().min(3).max(50),
     description: z.string().optional().nullable(),
-    duration: z.number().positive(),
-    price: z.number().positive(),
+    duration: z.number().positive().int(),
+    price: z.number().positive().int(),
 });
 
 const movieCreateSchema = z.object({
     name: z.string().min(3).max(50),
     description: z.string().optional().nullable(),
-    duration: z.number().positive(),
-    price: z.number().positive(),
+    duration: z.number().positive().int(),
+    price: z.number().positive().int(),
 });
 
 const returnMovieSchema = movieSchema.extend({
